+++ package.json
@@ -5,6 +5,7 @@
     "@dydxprotocol/v4-client-js": "^1.0.3",
     "@headlessui/react": "^0.0.0-insiders.bfacb64",
     "@headlessui/tailwindcss": "^0.2.0",
+    "@hookform/resolvers": "^3.3.2",
     "@ignt/react-library": "^0.4.3",
     "@tanstack/react-query": "^4.36.1",
     "@vercel/node": "^3.0.9",
@@ -26,10 +27,12 @@
     "qrcode": "^1.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.48.2",
     "react-inlinesvg": "^4.1.0",
     "react-router-dom": "^6.18.0",
     "use-debounce": "^9.0.4",
     "usehooks-ts": "^2.9.1",
+    "valibot": "^0.21.0",
     "viem": "^1.18.3",
     "vite-plugin-vercel-api": "^0.4.0",
     "wagmi": "^1.4.5",
