+++ src/components/dialog/EditRecordDialog.tsx
@@ -1,0 +1,157 @@
+import { DeliverTxResponse } from "@cosmjs/stargate";
+import { Dialog } from "@headlessui/react";
+import BaseDialog from "@/components/dialog/BaseDialog";
+import { useStore } from "@/store/index";
+import Button from "@/components/Button";
+import Radio, { Option } from "@/components/Radio";
+import { useClient } from "@/hooks/useClient";
+import { useState, useEffect } from "react";
+import { DnsRecordType } from "mycel-client-ts/mycel.registry/types/mycel/registry/dns_record";
+import { NetworkName } from "mycel-client-ts/mycel.registry/types/mycel/registry/network_name";
+import { RegistryRecord } from "mycel-client-ts/mycel.resolver/rest";
+import Dropdown from "@/components/Dropdown";
+import TxDialog from "@/components/dialog/TxDialog";
+import { Domain } from "@/types/domain";
+
+interface EditRecordDialogProps {
+  domain: Domain;
+  records: Record<string, RegistryRecord> | undefined;
+  address: string;
+}
+
+const recordOptions: Option[] = [
+  { value: "wallet", label: "Wallet Record" },
+  { value: "dns", label: "DNS Record" },
+];
+
+const RecordTypeToOptions = (recordType: any) => {
+  const options: Option[] = [];
+  Object.values(recordType).forEach((value) => {
+    if (typeof value === "string" && value !== "UNRECOGNIZED") {
+      options.push({ value: value, label: value });
+    }
+  });
+  return options;
+};
+
+export default function EditRecordDialog({ domain, records, address }: EditRecordDialogProps) {
+  const dialog = useStore((state) => state.dialog);
+  const updateDialog = useStore((state) => state.updateDialog);
+  const client = useClient();
+  const [recordOption, setRecordOption] = useState("wallet");
+  const [typeOption, setTypeOption] = useState("");
+  const [typeOptions, setTypeOptions] = useState<Option[]>([]);
+  const [currentRecordValue, setCurrentRecordValue] = useState("");
+  const [newRecordValue, setNewRecordValue] = useState("");
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [txResponse, setTxResponse] = useState<DeliverTxResponse>();
+
+  useEffect(() => {
+    if (recordOption === "dns") {
+      setTypeOptions(RecordTypeToOptions(DnsRecordType));
+      setTypeOption("A");
+    } else if (recordOption === "wallet") {
+      setTypeOptions(RecordTypeToOptions(NetworkName));
+      setTypeOption("ETHEREUM_TESTNET_GOERLI");
+    }
+  }, [recordOption]);
+
+  useEffect(() => {
+    if (records && records[typeOption] !== undefined) {
+      switch (recordOption) {
+        case "dns":
+          setCurrentRecordValue(records[typeOption].dnsRecord?.value || "");
+          break;
+        case "wallet":
+          setCurrentRecordValue(records[typeOption].walletRecord?.value || "");
+          break;
+      }
+    } else {
+      setCurrentRecordValue("");
+    }
+    setNewRecordValue("");
+  }, [typeOption]);
+
+  const handleRecordOptionChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setRecordOption(event.target.value);
+  };
+
+  const handleRecordTypeChange = (option: any) => {
+    setTypeOption(option.value);
+  };
+
+  // update record value
+  const updateRecord = async () => {
+    setIsLoading(true);
+    updateDialog("tx");
+    // update DNS record
+    if (recordOption === "dns") {
+      await client.MycelRegistry.tx
+        .sendMsgUpdateDnsRecord({
+          value: {
+            creator: address,
+            name: domain?.name || "",
+            parent: domain?.parent || "",
+            dnsRecordType: typeOption,
+            value: newRecordValue,
+          },
+        })
+        .then((res) => {
+          setIsLoading(false);
+          setTxResponse(res as DeliverTxResponse);
+        })
+        .catch(() => {
+          updateDialog(undefined);
+        });
+
+      // update wallet record
+    } else if (recordOption === "wallet") {
+      await client.MycelRegistry.tx
+        .sendMsgUpdateWalletRecord({
+          value: {
+            creator: address,
+            name: domain?.name || "",
+            parent: domain?.parent || "",
+            walletRecordType: typeOption,
+            value: newRecordValue,
+          },
+        })
+        .then((res) => {
+          setIsLoading(false);
+          setTxResponse(res as DeliverTxResponse);
+        })
+        .catch(() => {
+          updateDialog(undefined);
+        });
+    }
+  };
+
+  return (
+    <BaseDialog open={dialog === "editRecord"}>
+      <Dialog.Title className="text-2xl font-semibold mb-8 text-center">Edit Record</Dialog.Title>
+      <div>
+        <div className="relative flex-auto">
+          <Radio options={recordOptions} selectedOption={recordOption} onChange={handleRecordOptionChange} />
+          <div className="text-xs text-gray-600 pt-2">Record Type</div>
+          <Dropdown options={typeOptions} selectedOption={typeOption} onSelect={handleRecordTypeChange} />
+          <div className="text-xs text-gray-600 pt-2">CurrentRecord Record</div>
+          <h2 className="text-l font-semibold py-2">{currentRecordValue ? currentRecordValue : "---"}</h2>
+          <div className="text-xs text-gray-600 pt-2">New Record</div>
+          <input
+            className="mt-1 py-2 px-4 h-12 bg-white border border-black text-base leading-tight w-full outline-0"
+            value={newRecordValue}
+            onChange={(e) => setNewRecordValue(e.target.value)}
+          />
+          <Button
+            disabled={!address || newRecordValue === ""}
+            onClick={updateRecord}
+            className="btn-primary mt-10 h-10 w-48"
+          >
+            Update
+          </Button>
+        </div>
+        <TxDialog txResponse={txResponse} isLoading={isLoading} />
+      </div>
+    </BaseDialog>
+  );
+}
