+++ src/views/RegisterView.tsx
@@ -5,11 +5,12 @@ import useWallet from "@/hooks/useWallet";
 import { useMycelRegistry } from "@/hooks/useMycelRegistry";
 import { DeliverTxResponse } from "@cosmjs/stargate";
 import { PencilRuler } from "lucide-react";
-import TxModal from "@/components/TxModal";
+import TxDialog from "@/components/dialog/TxDialog";
 import ResolveButton from "@/components/ResolveButton";
 import { convertToDomain } from "@/utils/domainName";
 import { Domain } from "@/types/domain";
 import { MYCEL_COIN_DECIMALS, MYCEL_HUMAN_COIN_UNIT, convertToDecimalString } from "@/utils/coin";
+import { useStore } from "@/store/index";
 
 export default function RegisterView() {
   const { isConnected, mycelAccount } = useWallet();
@@ -18,10 +19,10 @@ export default function RegisterView() {
   const { secondLevelDomain, fee, registryQueryDomain, registryQueryRegistrationFee } = useMycelRegistry();
   const [query, setQuery] = useState<string>("");
   const [domain, setDomain] = useState<Domain>();
-  const [isShow, setIsShow] = useState<boolean>(false);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [txResponse, setTxResponse] = useState<DeliverTxResponse>();
   const [error, setError] = useState<string>("");
+  const updateDialog = useStore((state) => state.updateDialog);
 
   useEffect(() => {
     setDomain(convertToDomain(query));
@@ -37,7 +38,7 @@ export default function RegisterView() {
 
   const registerDomain = async () => {
     setIsLoading(true);
-    setIsShow(true);
+    updateDialog("tx");
 
     if (domain) {
       if (domain.parent === "") {
@@ -56,7 +57,7 @@ export default function RegisterView() {
           })
           .catch((error) => {
             setError(error.message);
-            setIsShow(false);
+            updateDialog(undefined);
           });
       } else {
         // Register second level domain
@@ -75,7 +76,7 @@ export default function RegisterView() {
           })
           .catch((error) => {
             setError(error.message);
-            setIsShow(false);
+            updateDialog(undefined);
           });
       }
     }
@@ -98,13 +99,12 @@ export default function RegisterView() {
             }}
           />
         </div>
-
         {fee && fee.isRegistrable ? (
           <div className="border-t border-b border-dashed border-black py-8 px-4">
             <div className="w-full flex justify-between">
-              <h2 className=" text-2xl m-2 font-semibold">{query}</h2>
+              <h2 className="text-2xl m-2 font-semibold">{query}</h2>
               {fee.fee && fee.fee[0].amount && (
-                <h2 className=" text-2xl pl-5 m-2">
+                <h2 className="text-2xl pl-5 m-2">
                   {convertToDecimalString(fee.fee[0].amount, MYCEL_COIN_DECIMALS)} {MYCEL_HUMAN_COIN_UNIT}/Year{" "}
                 </h2>
               )}
@@ -133,9 +133,8 @@ export default function RegisterView() {
         )}
         {error !== "" && <h2 className="  m-2 text-error font-semibold">{error}</h2>}
       </div>
-      <TxModal
-        isShow={isShow}
-        setIsShow={setIsShow}
+
+      <TxDialog
         txResponse={txResponse}
         isLoading={isLoading}
         onClosed={() => {
